Be/db/entity

See merge request s06-final/S06P31D204!8

+++ BE/src/main/java/com/marketganada/db/repository/TtestRepository.java
@@ -3,6 +3,6 @@ package com.marketganada.db.repository;
 import com.marketganada.db.entity.Ttest;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TtestRepository extends JpaRepository<Ttest,Integer>{
+public interface TtestRepository extends JpaRepository<Ttest,Long>{
 
 }
